requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pandas
openpyxl
python-multipart
selenium
webdriver-manager
python-dotenv